message_fetch: Move narrowing query builder to zerver.lib.narrow.
